Add PREMIUM_ROLE, refactor addProduct & deleteProduct methods

+++ src/controllers/product.controller.js
@@ -1,5 +1,6 @@
 import { ProductService } from '../repositories/index.js';
 import CustomError from '../utils/errors/custom.errors.js';
+import { logger } from '../utils/logger.js';
 
 export const getProducts = async (req, res) => {
     const { limit, page, sort, query } = req.query;
@@ -35,10 +36,23 @@ export const addProduct = async (req, res, next) => {
             CustomError.createProduct(data);
         }
 
+        const user = req.user;
+        
+        if(!user) {
+            throw new Error("User is null or undefined.");
+        }
+        
+        if(user.role !== "ADMIN_ROLE" && user.role !== "PREMIUM_ROLE") {
+            throw new Error("The user role is not allowed to add products.");
+        }
+
+        data.owner = user?.role === "ADMIN_ROLE" ? 'admin' : user.email;
+  
         const addProductRes = await ProductService.addProduct(data);
         return res.json({ productId: addProductRes });
     }
     catch(error) {
+        logger.debug(error)
         next(error);
     }
 };
@@ -69,10 +83,17 @@ export const deleteProduct = async (req, res, next) => {
             CustomError.deleteProduct();
         }
 
+        const pOwner = await ProductService.getProductOwner(pid);
+
+        if(pOwner.owner !== req.user.email && req.user.role !== "ADMIN_ROLE") {
+            throw new Error("The user does not have permission to delete this product."); 
+        }
+
         const deleteProductRes = await ProductService.deleteProduct(pid);
         return res.json({ productId: deleteProductRes });
     }
     catch(error) {
         next(error);
+        return;
     }
 };
